daemon: move basic `jsonrpc_` methods to an individual module
